+++ Bakeries.Web/bakerieswepapp/lib/screens/stock/widgets/add_stock_dialog/form_sections/quantity_section.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import '../../../../../constants/units.dart';
 
 class QuantitySection extends StatelessWidget {
-  final int quantity;
+  final double quantity;
   final String? selectedUnit;
-  final Function(int) onQuantityChanged;
+  final Function(double) onQuantityChanged;
   final Function(String) onUnitChanged;
 
   const QuantitySection({
@@ -24,7 +24,7 @@ class QuantitySection extends StatelessWidget {
             initialValue: quantity.toString(),
             keyboardType: TextInputType.number,
             decoration: const InputDecoration(labelText: 'الكمية'),
-            onChanged: (val) => onQuantityChanged(int.tryParse(val) ?? 0),
+            onChanged: (val) => onQuantityChanged(double.tryParse(val) ?? 0),
             validator: (value) {
               if (value?.isEmpty ?? true) return 'يجب إدخال الكمية';
               if (!RegExp(r'^\d+$').hasMatch(value!)) return 'أدخل رقم صحيح';
